separate api related codes to another folder of files
