+++ src/components/card-details.js
@@ -40,7 +40,7 @@ export default function DetailsPage() {
 
       {/* [forex] symbol: R_50, market: forex */}
       {/* [crypto] symbol: BTC, market: crypto */}
-      <Chart symbol="frxUSDSEK" market="forex" />
+      <Chart symbol="BTC" market="crypto" />
     </Card>
   );
 }
