+++ src/utils/utils.js
@@ -1,9 +1,3 @@
-const app_id = 1089;
-
-export const ws = new WebSocket(
-  "wss://ws.binaryws.com/websockets/v3?app_id=" + app_id
-);
-
 export const candleIntervals = {
   one_tick: { name: "one_tick", unit: "tick", value: 1, seconds: 1 },
   one_minute: { name: "one_minute", unit: "minute", value: 1, seconds: 60 },
@@ -170,38 +164,4 @@ export const createOHLC = (tick) => {
   newOHLC.open = newOHLC.close = newOHLC.high = newOHLC.low = tick.price;
 
   return newOHLC;
-};
-
-// close data stream
-export const closeStream = (stream_id) => {
-  console.log("closing ", typeof stream_id);
-  ws.send(
-    JSON.stringify({
-      forget: stream_id,
-    })
-  );
-};
-
-// get historical data
-export const getHistoricalData = (symbol, style, interval) => {
-  ws.send(
-    JSON.stringify({
-      ticks_history: symbol,
-      adjust_start_time: 1,
-      count: 100,
-      end: "latest",
-      style: style,
-      granularity: interval.seconds,
-    })
-  );
-};
-
-// get tick stream
-export const subscribeTickStream = (symbol) => {
-  ws.send(
-    JSON.stringify({
-      ticks: symbol,
-      subscribe: 1,
-    })
-  );
 };
